Extract transmission logic from Producer
